+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.4.1
+* Bugfix: Minor style adjustments for the horizontal bar
+
 ### 1.4.0
 * Enhancement: V9 Support and Internal code quality improvements (thanks to the community for the help!)
 * Bugfix: Fixed a specific situation when double clicking on minimize would double minimize